dry up command usage
