+++ db/migrate/20160407181257_remove_macros_command.rb
@@ -7,7 +7,7 @@ class RemoveMacrosCommand < ActiveRecord::Migration
     self.table_name = :macro_commands
   end
 
-  def change
+  def up
     OldMacro.find_each do |m|
       next unless command = m.command.presence
       max = OldMacroCommand.where(macro_id: m.id).maximum(:position) || 0
@@ -16,5 +16,9 @@ class RemoveMacrosCommand < ActiveRecord::Migration
     end
 
     remove_column :macros, :command
+  end
+
+  def down
+    add_column :macros, :command, :text, null: false, default: ""
   end
 end
